rename cannon to canon, remove id, set week to primary key
