+++ src/routes/index.ts
@@ -1,2 +1,2 @@
 export * from './auth/auth.routes';
-export * from './cannon/cannon.routes';
+export * from './canon/canon.routes';
